requirements.txt
numpy
pandas
PyYAML
requests
paho-mqtt
pluggy